Move App.cjx content into Dashboard.cjx
